requirements.txt
blinker==1.9.0
click==8.1.8
cobble==0.1.4
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
mammoth==1.9.0
MarkupSafe==3.0.2
olefile==0.47
pillow==11.1.0
PyPDF2==3.0.1
python-docx==1.1.2
python-magic==0.4.27
python-pptx==1.0.2
typing_extensions==4.12.2
Werkzeug==3.1.3
XlsxWriter==3.2.2
